added metadata to all routes
