+++ src/app/layout.tsx
@@ -7,13 +7,13 @@ import AuthWrapper from "@/auth_wrapper";
 
 const inter = Inter({ subsets: ["latin"] });
 
-// export const metadata: Metadata = {
-//   title: {
-//     template: '%s | Handcrafted Haven',
-//     default: 'Handcrafted Haven'
-//   },
-//   description: "Handcrafted Haven is an innovative web application that aims to provide a platform for artisans and crafters to showcase and sell their unique handcrafted items.",
-// };
+export const metadata: Metadata = {
+  title: {
+    template: '%s | Handcrafted Haven',
+    default: 'Handcrafted Haven'
+  },
+  description: "Handcrafted Haven is an innovative web application that aims to provide a platform for artisans and crafters to showcase and sell their unique handcrafted items.",
+};
 
 export default function LayoutContent({ children }: { children: ReactNode}) {
