+++ src/app/profile/create/page.tsx
@@ -1,5 +1,10 @@
 import CreateForm from "../../ui/profile/create_form";
 import { Suspense } from "react";
+import { Metadata } from 'next';
+
+export const metadata: Metadata = {
+  title: 'Create Product'
+}
 
 export default async function Page() {
   return (
