+++ src/app/register/page.tsx
@@ -1,116 +1,10 @@
-"use client";
+import { Metadata } from "next";
+import RegisterForm from "../ui/registerForm";
 
-import Link from "next/link";
-import { createUser, State } from "../lib/actions";
-import { useFormState } from "react-dom";
+export const metadata: Metadata = {
+  title: "Register",
+};
 
 export default function RegisterPage() {
-    const initialState: State = { message: null, errors: {} };
-    const [state, formAction] = useFormState(createUser, initialState);
-
-    return (
-        <main className="flex min-h-screen items-center justify-center bg-gray-100 p-6">
-            <div className="w-full max-w-md rounded-lg bg-white p-8 shadow-md">
-                <h2 className="mb-6 text-center text-2xl font-bold text-gray-800">Register</h2>
-                <form action={formAction} className="space-y-4">
-                    <div>
-                        <label htmlFor="firstName" className="block text-sm font-medium text-gray-700">
-                            First Name
-                        </label>
-                        <input
-                            id="firstName"
-                            name="firstName"
-                            type="text"
-                            placeholder="Please enter your first name"
-                            aria-describedby="firstName-error"
-                            className="mt-1 w-full rounded-md border border-gray-300 p-2 shadow-sm focus:border-indigo-500 focus:ring-indigo-500"
-                        />
-                        <div id="firstName-error" aria-live="polite" aria-atomic="true">
-                            {state.errors?.firstName &&
-                                state.errors.firstName.map((error: string) => (
-                                <p className="mt-2 text-sm text-red-500" key={error}>
-                                    {error}
-                                </p>
-                            ))}
-                        </div>
-                    </div>
-                    <div>
-                        <label htmlFor="lastName" className="block text-sm font-medium text-gray-700">
-                            Last Name
-                        </label>
-                        <input
-                            id="lastName"
-                            name="lastName"
-                            type="text"
-                            placeholder="Please enter your last name"
-                            aria-describedby="lastName-error"
-                            className="mt-1 w-full rounded-md border border-gray-300 p-2 shadow-sm focus:border-indigo-500 focus:ring-indigo-500"
-                        />
-                        <div id="lastName-error" aria-live="polite" aria-atomic="true">
-                            {state.errors?.lastName &&
-                                state.errors.lastName.map((error: string) => (
-                                <p className="mt-2 text-sm text-red-500" key={error}>
-                                    {error}
-                                </p>
-                            ))}
-                        </div>
-                    </div>
-                    <div>
-                        <label htmlFor="email" className="block text-sm font-medium text-gray-700">
-                            Email
-                        </label>
-                        <input
-                            id="email"
-                            name="email"
-                            type="email"
-                            placeholder="Please enter your email"
-                            aria-describedby="email-error"
-                            className="mt-1 w-full rounded-md border border-gray-300 p-2 shadow-sm focus:border-indigo-500 focus:ring-indigo-500"
-                        />
-                        <div id="email-error" aria-live="polite" aria-atomic="true">
-                            {state.errors?.email &&
-                                state.errors.email.map((error: string) => (
-                                <p className="mt-2 text-sm text-red-500" key={error}>
-                                    {error}
-                                </p>
-                            ))}
-                        </div>
-                    </div>
-                    <div>
-                        <label htmlFor="password" className="block text-sm font-medium text-gray-700">
-                            Password
-                        </label>
-                        <input
-                            id="password"
-                            name="password"
-                            type="password"
-                            placeholder="Please enter your password"
-                            aria-describedby="password-error"
-                            className="mt-1 w-full rounded-md border border-gray-300 p-2 shadow-sm focus:border-indigo-500 focus:ring-indigo-500"
-                        />
-                        <div id="password-error" aria-live="polite" aria-atomic="true">
-                            {state.errors?.password &&
-                                state.errors.password.map((error: string) => (
-                                <p className="mt-2 text-sm text-red-500" key={error}>
-                                    {error}
-                                </p>
-                            ))}
-                        </div>
-                    </div>
-                    <button
-                        type="submit"
-                        className="w-full rounded-lg bg-customGreen px-4 py-2 text-white transition-colors hover:bg-green-400"
-                    >
-                        Register
-                    </button>
-                </form>
-                <div className="mt-4 text-center text-sm text-gray-600">
-                    Already have an account?{" "}
-                    <Link href="/login" className="font-medium text-indigo-600 hover:text-indigo-500">
-                        Log in
-                    </Link>
-                </div>
-            </div>
-        </main>
-    );
+    return <RegisterForm />
 }
